Move auth+wake back to proxy as handle_client
